# Conflicts:
#	com.unity.render-pipelines.universal/Runtime/UniversalRenderPipeline.cs

+++ com.unity.render-pipelines.universal/Shaders/2D/Include/CombinedShapeLightShared.hlsl
@@ -14,7 +14,7 @@ half4 CombinedShapeLightShared(half4 color, half4 mask, half2 lightingUV)
 
     if (any(_ShapeLightMaskFilter0))
     {
-        float4 processedMask = (1 - _ShapeLightInvertedFilter0) * mask + _ShapeLightInvertedFilter0 * (1 - mask);
+        half4 processedMask = (1 - _ShapeLightInvertedFilter0) * mask + _ShapeLightInvertedFilter0 * (1 - mask);
         shapeLight0 *= dot(processedMask, _ShapeLightMaskFilter0);
     }
 
@@ -30,7 +30,7 @@ half4 CombinedShapeLightShared(half4 color, half4 mask, half2 lightingUV)
 
     if (any(_ShapeLightMaskFilter1))
     {
-        float4 processedMask = (1 - _ShapeLightInvertedFilter1) * mask + _ShapeLightInvertedFilter1 * (1 - mask);
+        half4 processedMask = (1 - _ShapeLightInvertedFilter1) * mask + _ShapeLightInvertedFilter1 * (1 - mask);
         shapeLight1 *= dot(processedMask, _ShapeLightMaskFilter1);
     }
 
@@ -46,7 +46,7 @@ half4 CombinedShapeLightShared(half4 color, half4 mask, half2 lightingUV)
 
     if (any(_ShapeLightMaskFilter2))
     {
-        float4 processedMask = (1 - _ShapeLightInvertedFilter2) * mask + _ShapeLightInvertedFilter2 * (1 - mask);
+        half4 processedMask = (1 - _ShapeLightInvertedFilter2) * mask + _ShapeLightInvertedFilter2 * (1 - mask);
         shapeLight2 *= dot(processedMask, _ShapeLightMaskFilter2);
     }
 
@@ -62,7 +62,7 @@ half4 CombinedShapeLightShared(half4 color, half4 mask, half2 lightingUV)
 
     if (any(_ShapeLightMaskFilter3))
     {
-        float4 processedMask = (1 - _ShapeLightInvertedFilter3) * mask + _ShapeLightInvertedFilter3 * (1 - mask);
+        half4 processedMask = (1 - _ShapeLightInvertedFilter3) * mask + _ShapeLightInvertedFilter3 * (1 - mask);
         shapeLight3 *= dot(processedMask, _ShapeLightMaskFilter3);
     }
